requirements.txt
# Core dependencies with pinned versions for reproducibility
# Last updated: 2024-02

# LLM and Embeddings
openai>=1.100.0,<2.0.0  # Require 1.100+ to avoid httpx proxies bug
tiktoken>=0.6.0
httpx>=0.24.0

# RAG Framework - Using latest stable LangChain
langchain==0.1.6
langchain-openai==0.0.6
langchain-community==0.0.19

# Vector Database
chromadb==0.4.22
onnxruntime>=1.17.0

# Search and Retrieval
rank-bm25==0.2.2
elasticsearch==8.12.0

# Reranking
cohere==4.47

# Evaluation
ragas==0.1.5
rouge-score==0.1.2
nltk==3.8.1
datasets==2.16.1

# NLI for Hallucination Detection
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Document Parsing
PyPDF2==3.0.1  # PDF parsing
python-docx==1.1.0  # DOCX parsing
pypdf==4.0.1  # Alternative PDF library with better text extraction

# UI
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Type hints
typing-extensions==4.9.0

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0

# Caching
redis==5.0.1

# Additional utilities
spacy==3.7.2
